eliminate some deps from base server
